Removal of all warnings
- Wykomentowanie w CmakeLists.txt wiersza "# SET(CMAKE_CXX_FLAGS
"-std=c++11 -pthread")"
- Usunięcie nieużywanych zmiennych
- Dodanie w pliku 3dsloader.cpp przed include'ami wiersza "#define
_CRT_SECURE_NO_DEPRECATE"
- Zostawienie w plikach "functions.h" i "functions.cpp" tylko dwóch
funkcji używanych przez klasę "objects3DS" (calcNormal i ReduceToUnit) -
reszta była nieużywana i generowały warningi
- W konstruktorze klasy CollisionDetectionColdet w pętli "for(int i=1;
i<nazwy_czesci.size(); i++)" definicja "int i" zamieniona na "size_t i",
ponieważ funkcja "size" zwraca typ "size_t", który jest unsigned, a int
jest signed.

+++ include/CollisionDetection/CollisionDetectionColdet.h
@@ -131,7 +131,7 @@ class CollisionDetectionColdet : public coldet::CollisionDetection {
 
 
                 std::cout << nazwy_czesci.front() << " length is: " << platform_length << " and width is: " << platform_width <<"\n";
-                for(int i=1; i<nazwy_czesci.size(); i++)
+                for(size_t i=1; i<nazwy_czesci.size(); i++)
                     std::cout << nazwy_czesci[i] << " length is: " << links_lengths[i-1] <<"\n";
                 std::cout << "Configuration done.\n";
 			}
